Moved mattermost from under email
